content/events/test/test_bug716822.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=716822
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 716822</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=716822">Mozilla Bug 716822</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">
/** Test for Bug 716822 **/
for (var p in window.NSEvent) {
  ok(p in window.Event, p + " should be in Event.");
  is(window.NSEvent[p], window.Event[p], "Value of " + p);
}
</script>
</pre>
</body>
</html>
